Reference

The following provides detailled documentation about types and methods provided by the LinearInterpolators package. This information is also available from the REPL by typing ? followed by the name of a method or a type.

Tabulated Interpolators

LinearInterpolators.Interpolations.TabulatedInterpolators.TabulatedInterpolator — Type

TabulatedInterpolator([T,] [d = nothing,] ker, pos, nrows, ncols)

yields a linear map to interpolate with kernel ker along a dimension of length ncols to produce a dimension of length nrows. The function pos(i) for i ∈ 1:nrows gives the positions to interpolate in fractional index unit along a dimension of length ncols. Argument d is the rank of the dimension along which to interpolate when the operator is applied to a multidimensional array. If it is unspecifed when the operator is created, it will have to be specified each time the operator is applied (see below).

The positions to interpolate can also be specified by a vector x as in:

TabulatedInterpolator([T,] [d = nothing,] ker, x, ncols)

to produce an interpolator whose output dimension is nrows = length(x). Here x can be an abstract vector.

Optional argument T is the floating-point type of the coefficients of the linear map. By default, it is given by the promotion of the element type of the arguments ker and, if specified, x.

The tabulated operator, say A, can be applied to an argument x:

apply!(α, P::Operations, A, [d,] x, scratch, β, y) -> y

to overwrite y with α*P(A)⋅x + β*y. If x and y are multi-dimensional, the dimension d to interpolate must be specified.

Two Dimensional Interpolators

LinearInterpolators.Interpolations.TwoDimensionalTransformInterpolator — Type

TwoDimensionalTransformInterpolator(rows, cols, ker1, ker2, R)

yields a linear mapping which interpolate its input of size cols to produce an output of size rows by 2-dimensional interpolation with kernels ker1 and ker2 along each dimension and applying the affine coordinate transform specified by R.

As a shortcut:

TwoDimensionalTransformInterpolator(rows, cols, ker, R)

is equivalent to TwoDimensionalTransformInterpolator(rows,cols,ker,ker,R) that is the same kernel is used along all dimensions.

Sparse Interpolators

LinearInterpolators.Interpolations.SparseInterpolators.SparseInterpolator — Type

A = SparseInterpolator{T=eltype(ker)}(ker, pos, grd)

yields a sparse linear interpolator suitable for interpolating with kernel ker a function sampled on the grid grd at positions pos. Optional parameter T is the floating-point type of the coefficients of the operator A. Call eltype(A) to query the type of the coefficients of the sparse interpolator A.

Then y = apply(A, x) or y = A(x) or y = A*x yields the interpolated values for interpolation weights x. The shape of y is the same as that of pos. Formally, this amounts to computing:

y[i] = sum_j ker((pos[i] - grd[j])/step(grd))*x[j]

with step(grd) the (constant) step size between the nodes of the grid grd and grd[j] the j-th position of the grid.

LinearInterpolators.Interpolations.fit — Function

fit(A, y [, w]; epsilon=1e-9, mu=0.0) -> x

performs a linear fit of y by the model A*x with A a linear interpolator. The returned value x minimizes:

sum(w.*(A*x - y).^2)

where w are given weights. If w is not specified, all weights are assumed to be equal to one; otherwise w must be an array of nonnegative values and of same size as y.

Keywords epsilon and mu may be specified to regularize the solution and minimize:

sum(w.*(A*x - y).^2) + rho*(epsilon*norm(x)^2 + mu*norm(D*x)^2)

where D is a finite difference operator, rho is the maximum diagonal element of A'*diag(w)*A and norm is the Euclidean norm.

LinearInterpolators.Interpolations.regularize — Function

regularize(A, ϵ, μ) -> R

regularizes the symmetric matrix A to produce the matrix:

R = A + ρ*(ϵ*I + μ*D'*D)

where I is the identity, D is a finite difference operator and ρ is the maximum diagonal element of A.

LinearInterpolators.Interpolations.regularize! — Function

regularize!(A, ϵ, μ) -> A

stores the regularized matrix in A (and returns it). This is the in-place version of [LinearInterpolators.SparseInterpolators.regularize].

LinearInterpolators.Interpolations.SparseInterpolators.SparseUnidimensionalInterpolator — Type

SparseUnidimensionalInterpolator{T=eltype(ker)}(ker, d, pos, grd)

yields a linear mapping which interpolates the d-th dimension of an array with kernel ker at positions pos along the dimension of interpolation d and assuming the input array has grid coordinates grd along the the d-th dimension of interpolation. Argument pos is a vector of positions, argument grd may be a range or the length of the dimension of interpolation. Optional parameter T is the floating-point type of the coefficients of the operator.

This kind of interpolator is suitable for separable multi-dimensional interpolation with precomputed interpolation coefficients. Having precomputed coefficients is mostly interesting when the operator is to be applied multiple times (for instance in iterative methods). Otherwise, separable operators which compute the coefficients on the fly may be preferable.

A combination of instances of SparseUnidimensionalInterpolator can be built to achieve sperable multi-dimensional interpolation. For example:

using LinearInterpolators
ker = CatmullRomSpline()
n1, n2 = 70, 50
x1 = linspace(1, 70, 201)
x2 = linspace(1, 50, 201)
A1 = SparseUnidimensionalInterpolator(ker, 1, x1, 1:n1)
A2 = SparseUnidimensionalInterpolator(ker, 2, x2, 1:n2)
A = A1*A2

Utilities

Limits

LinearInterpolators.Interpolations.Limits — Type

All interpolation limits inherit from the abstract type Limits{T} where T is the floating-point type. Interpolation limits are the combination of an extrapolation method and the length of the dimension to interpolate.

LinearInterpolators.Interpolations.limits — Function

limits(ker::Kernel, len)

yields the concrete type descendant of Limits for interpolation with kernel ker along a dimension of length len and applying the boundary conditions embedded in ker.

Interpolation coefficients

LinearInterpolators.Interpolations.getcoefs — Function

getcoefs(ker, lim, x) -> j1, j2, ..., w1, w2, ...

yields the indexes of the neighbors and the corresponding interpolation weights for interpolating at position x by kernel ker with the limits implemented by lim.